+++ fabric/operations.py
@@ -1127,7 +1127,8 @@ def local(command, capture=False, shell=None):
 
     In either case, as with `~fabric.operations.run` and
     `~fabric.operations.sudo`, this return value exhibits the ``return_code``,
-    ``stderr``, ``failed`` and ``succeeded`` attributes. See `run` for details.
+    ``stderr``, ``failed``, ``succeeded``, ``command`` and ``real_command``
+    attributes. See `run` for details.
 
     `~fabric.operations.local` will honor the `~fabric.context_managers.lcd`
     context manager, allowing you to control its current working directory
@@ -1140,6 +1141,8 @@ def local(command, capture=False, shell=None):
         Now honors the `~fabric.context_managers.lcd` context manager.
     .. versionchanged:: 1.0
         Changed the default value of ``capture`` from ``True`` to ``False``.
+    .. versionadded:: 1.9
+        The return value attributes ``.command`` and ``.real_command``.
     """
     given_command = command
     # Apply cd(), path() etc
@@ -1176,6 +1179,7 @@ def local(command, capture=False, shell=None):
     out = _AttributeString(stdout.strip() if stdout else "")
     err = _AttributeString(stderr.strip() if stderr else "")
     out.command = given_command
+    out.real_command = wrapped_command
     out.failed = False
     out.return_code = p.returncode
     out.stderr = err