+++ sites/www/changelog.rst
@@ -1,7 +1,9 @@
 =========
 Changelog
 =========
-
+* :feature:`1078` Add ``.command`` and ``.real_command`` attributes to
+  ``local`` return value.  Thanks to Alexander Teves (``@alexanderteves``) and
+  Konrad Hałas (``@konradhalas``).
 * :feature:`1101` Reboot operation now supports custom command. Thanks to Jonas
   Lejon.
 * :support:`1106` Fix a misleading/ambiguous example snippet in the ``fab``
@@ -19,7 +21,7 @@ Changelog
 * :bug:`955` Quote directories created as part of ``put``'s recursive directory
   uploads when ``use_sudo=True`` so directories with shell meta-characters
   (such as spaces) work correctly. Thanks to John Harris for the catch.
-* :bug:`917` Correct an issue with ``put(use_sudo=True, mode=xxx)`` where the 
+* :bug:`917` Correct an issue with ``put(use_sudo=True, mode=xxx)`` where the
   ``chmod`` was trying to apply to the wrong location. Thanks to Remco
   (``@nl5887``) for catch & patch.
 * :bug:`1046` Fix typo preventing use of ProxyCommand in some situations.